legacy/requirements.txt
# Local inferece
transformers~=4.48.0
torch~=2.5.1
bitsandbytes~=0.45.0
accelerate>=0.26.0

# For crawling
gitpython~=3.1.44

# Graph view
matplotlib
beautifulsoup4~=4.12.3
markdown~=3.7

# Interaction with Ollama
llama_index~=0.12.11
llama-index-llms-ollama~=0.5.0
llama-index-embeddings-ollama~=0.5.0
ollama~=0.4.6

# Store
manticoresearch-dev
pydantic_settings~=2.7.1
